# api/src/config/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PulseOps V2 API",
    "version": "2.0.0",
    "description": "Enterprise platform API for PulseOps V2. Supports database management, authentication (JWT + HttpOnly cookies), module hot-dropping, and system configuration. All URLs are defined in urls.json — this spec references relative paths under the /api prefix.",
    "contact": {
      "name": "PulseOps Team"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Relative to host — works on any environment"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained via POST /auth/login. Expiry: 24h."
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "HttpOnly cookie set automatically on login. Same JWT as BearerAuth."
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "description": "Standard success envelope returned by all endpoints on success.",
        "properties": {
          "success": { "type": "boolean", "example": true, "description": "Always true on success" },
          "data": { "type": "object", "description": "Endpoint-specific response payload" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error envelope returned by all endpoints on failure.",
        "properties": {
          "success": { "type": "boolean", "example": false, "description": "Always false on error" },
          "error": {
            "type": "object",
            "properties": {
              "message": { "type": "string", "description": "Human-readable error message from APIErrors.json", "example": "Invalid credentials" },
              "code": { "type": "string", "description": "Machine-readable error code", "example": "TOKEN_EXPIRED" }
            }
          }
        }
      },
...
      "DatabaseConfig": {
        "type": "object",
        "description": "PostgreSQL connection parameters. All fields optional — defaults from DatabaseConfig.json used if omitted.",
        "properties": {
          "host": { "type": "string", "description": "Database server hostname", "example": "db.example.com", "default": "localhost" },
          "port": { "type": "integer", "description": "Database server port (1-65535)", "example": 5432, "default": 5432 },
          "database": { "type": "string", "description": "Target database name", "example": "pulseops_v2", "default": "pulseops_v2" },
          "schema": { "type": "string", "description": "Schema name for all PulseOps tables", "example": "pulseops", "default": "pulseops" },
          "username": { "type": "string", "description": "Database user", "example": "postgres", "default": "postgres" },
          "password": { "type": "string", "format": "password", "description": "Database password" }
        }
      },
      "ConnectionTestResult": {
        "type": "object",
        "description": "Result of a database connection test.",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": {
            "type": "object",
            "properties": {
              "connected": { "type": "boolean", "description": "Whether connection succeeded", "example": true },
              "message": { "type": "string", "example": "Database connection successful" },
              "latencyMs": { "type": "integer", "description": "Connection latency in milliseconds", "example": 12 },
              "dbVersion": { "type": "string", "description": "PostgreSQL server version string", "example": "PostgreSQL 16.1" }
            }
          }
        }
      },
...
      "SchemaCreateResult": {
        "type": "object",
        "description": "Result of schema creation.",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": {
            "type": "object",
            "properties": {
              "message": { "type": "string", "example": "Database schema created successfully" },
              "tables": {
                "type": "array",
                "items": { "type": "string" },
                "example": ["system_users", "system_config", "system_modules", "system_logs"]
              }
            }
          }
        }
      },
      "DefaultDataResult": {
        "type": "object",
        "description": "Result of loading or cleaning default data.",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": {
            "type": "object",
            "properties": {
              "message": { "type": "string", "example": "Default data loaded successfully" },
              "usersInserted": { "type": "integer", "description": "Number of users seeded", "example": 1 },
              "modulesInserted": { "type": "integer", "description": "Number of modules registered", "example": 2 }
            }
          }
        }
      },
      "DatabaseStatsResult": {
        "type": "object",
        "description": "Table sizes and row counts.",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": {
            "type": "object",
            "properties": {
              "tables": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "name": { "type": "string", "example": "system_users" },
                    "rows": { "type": "integer", "example": 1 },
                    "size": { "type": "string", "example": "8192 bytes" }
                  }
                }
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/database/test-connection": {
      "get": {
        "tags": ["Database"],
        "operationId": "testConnectionServer",
        "summary": "Test connection with server config",
        "description": "Tests PostgreSQL connectivity using the server's DatabaseConfig.json settings. Returns connection status, latency in ms, and PostgreSQL version string.",
        "responses": {
          "200": { "description": "Connection test result.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConnectionTestResult" } } } }
        }
      },
      "post": {
        "tags": ["Database"],
        "operationId": "testConnectionCustom",
        "summary": "Test connection with custom config",
        "description": "Tests PostgreSQL connectivity using the provided configuration. Creates a temporary connection pool, runs SELECT version(), and returns latency. Pool is destroyed after test.",
        "requestBody": {
          "required": true,
          "description": "Database connection parameters to test.",
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DatabaseConfig" } } }
        },
        "responses": {
          "200": { "description": "Connection test result.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConnectionTestResult" } } } }
        }
      }
    },
    "/database/save-config": {
      "post": {
        "tags": ["Database"],
        "operationId": "saveDbConfig",
        "summary": "Save database configuration",
        "description": "Persists the provided database configuration to DatabaseConfig.json on the API server's filesystem. Takes effect on next connection pool creation.",
        "requestBody": {
          "required": true,
          "description": "Database configuration to save.",
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DatabaseConfig" } } }
        },
        "responses": {
          "200": { "description": "Config saved to DatabaseConfig.json.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "500": { "description": "File write failed.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/database/create-database": {
      "post": {
        "tags": ["Database"],
        "operationId": "createDatabase",
        "summary": "Create the database",
        "description": "Connects to the 'postgres' system database and runs CREATE DATABASE for the target database name from config. Returns success if database already exists.",
        "responses": {
          "200": { "description": "Database created or already exists.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "500": { "description": "Database creation failed.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/database/delete-database": {
      "delete": {
        "tags": ["Database"],
        "operationId": "deleteDatabase",
        "summary": "Delete the database",
        "description": "DESTRUCTIVE: Drops the entire PostgreSQL database including all schemas, tables, and data. Requires JWT authentication. Terminates all active connections before dropping.",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Database dropped.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "401": { "description": "Authentication required.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "500": { "description": "Drop failed.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
    "/database/create-schema": {
      "post": {
        "tags": ["Database"],
        "operationId": "createSchema",
        "summary": "Create database schema",
        "description": "Creates the core schema and 4 tables inside a transaction: system_users (auth), system_config (JSONB key-value), system_modules (hot-drop registry), system_logs (structured logs). Rolls back on any failure.",
        "responses": {
          "200": { "description": "Schema and tables created.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SchemaCreateResult" } } } },
          "500": { "description": "Schema creation failed (rolled back).", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/database/load-default-data": {
      "post": {
        "tags": ["Database"],
        "operationId": "loadDefaultData",
        "summary": "Load default data",
        "description": "Seeds the default admin user (from DefaultAdminUser.json with bcrypt-hashed password) and registers core modules (platform_admin, auth). Skips if data already exists.",
        "responses": {
          "200": { "description": "Default data loaded.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DefaultDataResult" } } } },
          "500": { "description": "Seeding failed.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "delete": {
        "tags": ["Database"],
        "operationId": "cleanDefaultData",
        "summary": "Clean default data",
        "description": "Removes all seeded default data: deletes the default admin user and unregisters core modules from the database.",
        "responses": {
          "200": { "description": "Default data removed.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } }
        }
      }
    },
    "/database/wipe": {
      "post": {
        "tags": ["Database"],
        "operationId": "wipeDatabase",
        "summary": "Wipe database schema",
        "description": "DESTRUCTIVE: Drops the entire schema CASCADE, removing all tables and data. Requires JWT authentication. The database itself is not dropped — only the schema.",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Schema wiped.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "401": { "description": "Authentication required.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/database/stats": {
      "get": {
        "tags": ["Database"],
        "operationId": "getDatabaseStats",
        "summary": "Get database stats",
        "description": "Returns table names, row counts, and disk sizes for all tables in the schema. Requires JWT authentication.",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Database statistics.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DatabaseStatsResult" } } } },
          "401": { "description": "Authentication required.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
  }
}